backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
passlib==1.7.4
redis==5.0.7
Werkzeug==3.0.3
